+++ src/load_champion.rs
@@ -117,7 +117,7 @@ pub fn load_champion_stats(champion_name: &str) -> ChampionStats {
 }
 
 impl ChampionStats {
-    pub fn as_basic_attack(&self, level: u32) -> BasicAttack {
+    pub fn as_basic_attack(&self, level: u8) -> BasicAttack {
         let attack_damage =
             core::stat_at_level(self.attack_damage, self.attack_damage_per_level, level);
         let critical_strike_chance =
@@ -131,7 +131,7 @@ impl ChampionStats {
         return attack;
     }
 
-    pub fn as_attack_speed(&self, level: u32) -> AttackSpeed {
+    pub fn as_attack_speed(&self, level: u8) -> AttackSpeed {
         let bonus_speed =
             core::stat_at_level(0.0, self.attack_speed_per_level, level) + self.bonus_attack_speed;
         return AttackSpeed {
@@ -140,7 +140,7 @@ impl ChampionStats {
         };
     }
 
-    pub fn as_target(&self, level: u32) -> Target {
+    pub fn as_target(&self, level: u8) -> Target {
         let base_armor = core::stat_at_level(self.armor, self.armor_per_level, level);
         let magic_resist =
             core::stat_at_level(self.magic_resist, self.magic_resist_per_level, level);
